refactor MyAppCompatActivity.kt into /config
