+++ app/src/main/java/com/github/fribourgsdp/radio/ImportSpotifyPlaylistsActivity.kt
@@ -3,6 +3,7 @@ package com.github.fribourgsdp.radio
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import com.github.fribourgsdp.radio.config.MyAppCompatActivity
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
 import com.github.fribourgsdp.radio.data.User
