+++ app/src/main/java/com/github/fribourgsdp/radio/JoinGameActivity.kt
@@ -13,6 +13,7 @@ import android.widget.TextView
 import androidx.core.content.ContextCompat
 import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.DialogFragment
+import com.github.fribourgsdp.radio.config.MyAppCompatActivity
 import com.github.fribourgsdp.radio.data.User
 
 const val GAME_UID_KEY = "com.github.fribourgsdp.radio.GAME_UID"
