+++ app/src/main/java/com/github/fribourgsdp/radio/data/view/PlaylistsFragmentHolderActivity.kt
@@ -1,7 +1,7 @@
 package com.github.fribourgsdp.radio.data.view
 
 import android.os.Bundle
-import com.github.fribourgsdp.radio.MyAppCompatActivity
+import com.github.fribourgsdp.radio.config.MyAppCompatActivity
 import com.github.fribourgsdp.radio.util.MyFragment
 import com.github.fribourgsdp.radio.R
