+++ app/src/main/java/com/github/fribourgsdp/radio/data/view/UserProfileActivity.kt
@@ -9,6 +9,7 @@ import android.widget.EditText
 import android.widget.ImageView
 import android.widget.TextView
 import com.github.fribourgsdp.radio.*
+import com.github.fribourgsdp.radio.config.MyAppCompatActivity
 import com.github.fribourgsdp.radio.data.User
 import com.github.fribourgsdp.radio.util.MyFragment
 import com.google.android.material.floatingactionbutton.FloatingActionButton
